issue-98: Add splash screen to updated rmain
- Update branch to be up to date with main
- Move splash_screen function to rmain.c
- Remove input.c include in splash.h
